# Make GSL as a Blas/Lapack 3rd party as well as MKL, OpenBLAS, Netlib, Atlas