… 3rdparty/clapack/CMakeLists.txt
… 3rdparty/clapack/blas/dsyr2k.c
… 3rdparty/clapack/src/dlansy.c
… 3rdparty/clapack/src/dlatrd.c
… 3rdparty/clapack/src/dorg2l.c
… 3rdparty/clapack/src/dorgql.c
… 3rdparty/clapack/src/dorgtr.c
… 3rdparty/clapack/src/dsteqr.c
… 3rdparty/clapack/src/dsterf.c
… 3rdparty/clapack/src/dsyev.c
… 3rdparty/clapack/src/dsytd2.c
… 3rdparty/clapack/src/dsytrd.c
… CMakeLists.txt
… cmake/ChooseBlas.cmake
… cmake/FindLAPACK_C.cmake
… cmake/FindNetlib.cmake
… cmake/VISPUtils.cmake
… cmake/templates/VISPConfig.cmake.in
… cmake/templates/vpConfig.h.in
… doc/mainpage.doc.in
… doc/tutorial/linear-algebra/logs/perfMatrixMultiplication/atlas-benchmark.txt
… doc/tutorial/linear-algebra/logs/perfMatrixMultiplication/built-in-benchmark.txt
… doc/tutorial/linear-algebra/logs/perfMatrixMultiplication/gsl-benchmark.txt
… doc/tutorial/linear-algebra/logs/perfMatrixMultiplication/mkl-benchmark.txt
… doc/tutorial/linear-algebra/logs/perfMatrixMultiplication/netlib-benchmark.txt
… doc/tutorial/linear-algebra/logs/perfMatrixMultiplication/openblas-benchmark.txt
… doc/tutorial/linear-algebra/logs/perfMatrixMultiplication/without-lapack-benchmark.txt
… doc/tutorial/linear-algebra/tutorial-basic-linear-algebra.doc
… doc/tutorial/tutorial.doc
… example/math/quadprog.cpp