Changes that were done are resumed hereafter:

- Make GSL usage as a Blas/Lapack 3rd party like MKL, OpenBLAS, Netlib, Atlas (Closes #735)
    - cmake option USE_BLAS/LAPACK allows now to select one 3rd party between
      [ MKL | OpenBLAS | Netlib | Atlas | GSL ]
    - computing SVD when number of rows < number of cols is now possible using GSL
    - extend existing functionalities for all the Lapack 3rd parties
      (a good example is vpMatrix::inverseTriangular() that was not
      implemented for MKL and GSL)
    - make all the vpMatrix functions suffixed with Gsl as deprecated
      (svdGsl(), pseudoInverseGsl()...). For compat with previous releases
      they are keept but excluded from Doxygen documentation
    - in vpConfig.h define VISP_HAVE_LAPACk_MKL in addition of VISP_HAVE_MKL
      to be more explicit. Same for the other Lapack 3rd parties
- Introduce new tutorial for basic linear algebra with Lapack usage
    - shows how to install one of the supported Blas/Lapack 3rd party under Ubuntu
    - compares performances obtained on some choosed basic linear algebra operations
    - save also the benchmarking results obtained with perfMatrixMultiplication to
      have a reference for future optimisation
- Various optimisations using Blas/Lapack or SSE2 in vpMatrix
    - optimise AAt() using Blas/Lapack
    - optimize operator*(const vpForceTwistMatrix &) using SSE2
    - introduce operator*(const vpHomogeneousMatrix &)
    - introduce getLapackMatrixMinSize() and setLapackMatrixMinSize() to
      be able to modify the matrix/vector default min size requested to use
      Lapack for optimization. By defaut, when one of the Lapack 3rd party is available
      Lapack is used