@s-trinh 2020-04-05
## Test with Intel MKL 2019.1.144 and Ubuntu 16.04

`Intel(R) Core(TM) i5-8500 CPU @ 3.00GHz`

### ./testGenericTracker -d -D
- Before:

```
Computation time, Mean: 28.0571 ms ; Median: 26.8651 ms ; Std: 10.0519 ms
Max translation error: 0.00110701
Max thetau error: 0.372656

Computation time, Mean: 27.987 ms ; Median: 26.6885 ms ; Std: 10.0708 ms
Max translation error: 0.00110701
Max thetau error: 0.372656

Computation time, Mean: 27.9939 ms ; Median: 26.8212 ms ; Std: 10.0406 ms
Max translation error: 0.00110701
Max thetau error: 0.372656
```

- After:

```
Computation time, Mean: 34.196 ms ; Median: 32.1661 ms ; Std: 14.2505 ms
Max translation error: 0.00110701
Max thetau error: 0.372656

Computation time, Mean: 33.605 ms ; Median: 31.7834 ms ; Std: 12.7761 ms
Max translation error: 0.00110701
Max thetau error: 0.372656

Computation time, Mean: 33.0729 ms ; Median: 31.5372 ms ; Std: 12.2484 ms
Max translation error: 0.00110701
Max thetau error: 0.372656
```

- Results:
  - small perf degradation

### ./testGenericTracker -d -D (5 back and forth tracking)
The program is modified to run the tracker on the same sequence back and forth 5 times to increase the running time of the program (400 frames instead of 40).

- Before:

```
Computation time, Mean: 27.4078 ms ; Median: 26.2208 ms ; Std: 9.76099 ms
Max translation error: 0.00111751
Max thetau error: 0.372656
Features, Mean: 51405.4 ; Median: 46066 ; Std: 13883.1
```

- After:

```
Computation time, Mean: 30.6809 ms ; Median: 29.6366 ms ; Std: 11.0347 ms
Max translation error: 0.00111751
Max thetau error: 0.372656
Features, Mean: 51404.7 ; Median: 46066 ; Std: 13881.4
```

- Results:
  - small perf degradation is confirmed, see `vpMatrix vpMatrix::operator*(const vpVelocityTwistMatrix &V) const` change
  - not the same mean number of features, can be explained since floating point operations are not computed in the same order between BLAS and SSE2 code
  - [proof](http://cpp.sh/8wm4b): 
```
// Example program
#include <iostream>
#include <string>
#include <iomanip>
#include <limits>

int main()
{
    double val1 = 0.1 + 0.2 + 0.3;
    double val2 = 0.1 + (0.2 + 0.3);
    
    std::cout << "val1: " << std::setprecision(std::numeric_limits<double>::max_digits10) << val1 << std::endl;
    std::cout << "val2: " << std::setprecision(std::numeric_limits<double>::max_digits10) << val2 << std::endl;
}
```
  - see also: [What does gcc's ffast-math actually do?](https://stackoverflow.com/questions/7420665/what-does-gccs-ffast-math-actually-do)

### ./perfMatrixMultiplication --benchmark

- Before:

<details><summary>log</summary>
<p>

```
./perfMatrixMultiplication --benchmark

~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
perfMatrixMultiplication is a Catch v2.9.2 host application.
Run with -? for options

-------------------------------------------------------------------------------
Benchmark matrix-matrix multiplication
-------------------------------------------------------------------------------
ViSP/ViSP-src/modules/core/test/math/perfMatrixMultiplication.cpp:217
...............................................................................

benchmark name                                  samples       iterations    estimated
                                                mean          low mean      high mean
                                                std dev       low std dev   high std dev
-------------------------------------------------------------------------------
(3x3)x(3x3) - Naive code                                100          152    2.1128 ms 
                                                     136 ns       131 ns       148 ns 
                                                      33 ns         1 ns        66 ns 
                                                                                      
(3x3)x(3x3) - ViSP                                      100           72    2.1168 ms 
                                                     298 ns       285 ns       331 ns 
                                                      97 ns        29 ns       179 ns 
                                                                                      
(6x6)x(6x6) - Naive code                                100           63    2.0979 ms 
                                                     333 ns       318 ns       371 ns 
                                                     105 ns         3 ns       202 ns 
                                                                                      
(6x6)x(6x6) - ViSP                                      100           76    2.0976 ms 
                                                     277 ns       276 ns       281 ns 
                                                       8 ns         0 ns        20 ns 
                                                                                      
(8x8)x(8x8) - Naive code                                100           44    2.1428 ms 
                                                     486 ns       486 ns       489 ns 
                                                       6 ns         0 ns        12 ns 
                                                                                      
(8x8)x(8x8) - ViSP                                      100           62    2.1018 ms 
                                                     339 ns       337 ns       345 ns 
                                                      14 ns         0 ns        34 ns 
                                                                                      
(10x10)x(10x10) - Naive code                            100           28    2.1252 ms 
                                                     856 ns       854 ns       860 ns 
                                                      12 ns         1 ns        25 ns 
                                                                                      
(10x10)x(10x10) - ViSP                                  100           57    2.1261 ms 
                                                     464 ns       462 ns       472 ns 
                                                      16 ns         1 ns        39 ns 
                                                                                      
(20x20)x(20x20) - Naive code                            100            4    2.1612 ms 
                                                    5.41 us     5.404 us     5.438 us 
                                                      57 ns         7 ns       135 ns 
                                                                                      
(20x20)x(20x20) - ViSP                                  100           21     2.184 ms 
                                                   1.061 us     1.055 us     1.087 us 
                                                      52 ns         2 ns       125 ns 
                                                                                      
(6x200)x(200x6) - Naive code                            100            3    2.1174 ms 
                                                   7.048 us     7.041 us     7.079 us 
                                                      63 ns         5 ns       150 ns 
                                                                                      
(6x200)x(200x6) - ViSP                                  100           33    2.1516 ms 
                                                     658 ns       653 ns       682 ns 
                                                      48 ns         2 ns       115 ns 
                                                                                      
(200x6)x(6x200) - Naive code                            100            1   15.2522 ms 
                                                 152.326 us   151.987 us   153.071 us 
                                                   2.448 us     1.202 us     4.057 us 
                                                                                      
(200x6)x(6x200) - ViSP                                  100            1     2.399 ms 
                                                  24.237 us    23.894 us    25.394 us 
                                                   2.883 us     1.046 us     6.461 us 
                                                                                      
(207x119)x(119x207) - Naive code                        100            1   470.526 ms 
                                                 4.70783 ms   4.70179 ms   4.71718 ms 
                                                  37.773 us     27.11 us    51.962 us 
                                                                                      
(207x119)x(119x207) - ViSP                              100            1   20.1039 ms 
                                                 201.249 us   200.787 us   203.211 us 
                                                   4.115 us       718 ns     9.611 us 
                                                                                      
(83x201)x(201x83) - Naive code                          100            1   133.006 ms 
                                                 1.33239 ms   1.32951 ms   1.33834 ms 
                                                  20.185 us    12.204 us    36.834 us 
                                                                                      
(83x201)x(201x83) - ViSP                                100            1    6.1966 ms 
                                                  62.423 us    62.093 us    63.256 us 
                                                     2.5 us     1.154 us     4.647 us 
                                                                                      
(600x400)x(400x600) - Naive code                        100            1    14.8022 s 
                                                 146.127 ms   146.095 ms   146.171 ms 
                                                 190.754 us   146.263 us   268.794 us 
                                                                                      
(600x400)x(400x600) - ViSP                              100            1   519.368 ms 
                                                 5.19594 ms   5.18278 ms   5.21372 ms 
                                                  77.287 us    61.879 us   100.503 us 
                                                                                      
(400x600)x(600x400) - Naive code                        100            1    9.96885 s 
                                                  99.665 ms   99.6371 ms   99.6976 ms 
                                                 153.093 us   131.092 us   184.946 us 
                                                                                      
(400x600)x(600x400) - ViSP                              100            1    337.41 ms 
                                                 3.37732 ms   3.36913 ms   3.38891 ms 
                                                      49 us    37.332 us    70.687 us 
                                                                                      

-------------------------------------------------------------------------------
Benchmark matrix-rotation matrix multiplication
-------------------------------------------------------------------------------
ViSP/ViSP-src/modules/core/test/math/perfMatrixMultiplication.cpp:294
...............................................................................

benchmark name                                  samples       iterations    estimated
                                                mean          low mean      high mean
                                                std dev       low std dev   high std dev
-------------------------------------------------------------------------------
(3x3)x(3x3) - Naive code                                100           73    2.1243 ms 
                                                     289 ns       289 ns       289 ns 
                                                       1 ns         0 ns         2 ns 
                                                                                      
(3x3)x(3x3) - ViSP                                      100          107    2.1186 ms 
                                                     197 ns       197 ns       198 ns 
                                                       1 ns         0 ns         3 ns 
                                                                                      

-------------------------------------------------------------------------------
Benchmark matrix-homogeneous matrix multiplication
-------------------------------------------------------------------------------
ViSP/ViSP-src/modules/core/test/math/perfMatrixMultiplication.cpp:379
...............................................................................

benchmark name                                  samples       iterations    estimated
                                                mean          low mean      high mean
                                                std dev       low std dev   high std dev
-------------------------------------------------------------------------------
(4x4)x(4x4) - Naive code                                100           61    2.1289 ms 
                                                     332 ns       331 ns       334 ns 
                                                       5 ns         0 ns        12 ns 
                                                                                      
(4x4)x(4x4) - ViSP                                      100           54     2.133 ms 
                                                     395 ns       394 ns       400 ns 
                                                      10 ns         0 ns        22 ns 
                                                                                      

-------------------------------------------------------------------------------
Benchmark matrix-vector multiplication
-------------------------------------------------------------------------------
ViSP/ViSP-src/modules/core/test/math/perfMatrixMultiplication.cpp:467
...............................................................................

benchmark name                                  samples       iterations    estimated
                                                mean          low mean      high mean
                                                std dev       low std dev   high std dev
-------------------------------------------------------------------------------
(3x3)x(3x1) - Naive code                                100          108    2.0952 ms 
                                                     194 ns       194 ns       194 ns 
                                                       1 ns         0 ns         2 ns 
                                                                                      
(3x3)x(3x1) - ViSP                                      100           97    2.0952 ms 
                                                     216 ns       216 ns       218 ns 
                                                       3 ns         0 ns         7 ns 
                                                                                      
(6x6)x(6x1) - Naive code                                100           98    2.0972 ms 
                                                     214 ns       214 ns       214 ns 
                                                       1 ns         0 ns         3 ns 
                                                                                      
(6x6)x(6x1) - ViSP                                      100           92     2.116 ms 
                                                     229 ns       229 ns       231 ns 
                                                       4 ns         0 ns        10 ns 
                                                                                      
(8x8)x(8x1) - Naive code                                100           93    2.1204 ms 
                                                     228 ns       228 ns       228 ns 
                                                       0 ns         0 ns         1 ns 
                                                                                      
(8x8)x(8x1) - ViSP                                      100           90     2.115 ms 
                                                     232 ns       232 ns       234 ns 
                                                       4 ns         0 ns        11 ns 
                                                                                      
(10x10)x(10x1) - Naive code                             100           85    2.1165 ms 
                                                     249 ns       248 ns       249 ns 
                                                       2 ns         0 ns         4 ns 
                                                                                      
(10x10)x(10x1) - ViSP                                   100           88     2.112 ms 
                                                     244 ns       240 ns       253 ns 
                                                      25 ns         4 ns        47 ns 
                                                                                      
(20x20)x(20x1) - Naive code                             100           45     2.133 ms 
                                                     474 ns       473 ns       476 ns 
                                                       6 ns         1 ns        11 ns 
                                                                                      
(20x20)x(20x1) - ViSP                                   100           66    2.1252 ms 
                                                     324 ns       322 ns       329 ns 
                                                      10 ns         0 ns        24 ns 
                                                                                      
(6x200)x(200x1) - Naive code                            100           22     2.156 ms 
                                                     984 ns       983 ns       988 ns 
                                                       9 ns         3 ns        19 ns 
                                                                                      
(6x200)x(200x1) - ViSP                                  100           71    2.1087 ms 
                                                     306 ns       305 ns       313 ns 
                                                      14 ns         1 ns        33 ns 
                                                                                      
(200x6)x(6x1) - Naive code                              100           19    2.1603 ms 
                                                    1.14 us     1.138 us      1.15 us 
                                                      20 ns         2 ns        47 ns 
                                                                                      
(200x6)x(6x1) - ViSP                                    100           27    2.1411 ms 
                                                     792 ns       788 ns       812 ns 
                                                      40 ns         2 ns        94 ns 
                                                                                      
(207x119)x(119x1) - Naive code                          100            2     3.262 ms 
…